requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.23
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
requests==2.26.0
aiohttp==3.8.1
redis==3.5.3
cryptography==3.4.8
pyjwt==2.1.0
urllib3==1.26.6 